Refactor code to replace int with uint32_t
The revisions replace instances of int with uint32_t to ensure consistency and limit the usage of conflicting data types. This adjustment provides better support for unsigned 32-bit integers across multiple code files, including AstNode, TypeMatch, and Generic. Several function parameters have been updated accordingly.

+++ src/ByteCodeGen_Function.cpp
@@ -1755,7 +1755,7 @@ bool ByteCodeGen::emitCall(ByteCodeGenContext* context,
             for (int j = 0; j < numCallParams; j++)
             {
                 auto param = castAst<AstFuncCallParam>(allParams->children[j], AstNodeKind::FuncCallParam);
-                if (param->indexParam == i)
+                if (param->indexParam == static_cast<uint32_t>(i))
                 {
                     if (param->hasExtMisc() && !param->extMisc()->additionalRegisterRC.cannotFree)
                     {
